Backend/pyproject.toml
[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
djangorestframework-stubs = "^3.15.3"
black = "^25.1.0"
django-stubs = "^5.1.3"
django-stubs-ext = "^5.1.3"

[project]
name = "filmnoir"
version = "0.1.0"
description = ""
authors = [
    { name = "nauzybekturganbayev", email = "smiling.cornflower@gmail.com" }
]
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "unidecode (>=1.3.8,<2.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0) ; python_version >= \"3.12\" and python_version < \"4.0\"",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",

]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]
# explicit_package_bases = true

warn_unused_ignores = true
follow_imports_for_stubs = true
ignore_missing_imports = true
show_absolute_path = true
hide_error_codes = false
show_error_context = true
strict = true
warn_unreachable = true
warn_no_return = true
cache_dir = ".mypy_cache"
exclude = "^migrations/.*"

[tool.django-stubs]
django_settings_module = "config.settings"
